@@ -15,12 +15,15 @@
     "ejs": "2.5.5"
   },
   "devDependencies": {
+    "babel-eslint": "^10.0.1",
     "chai": "^3.5.0",
     "chai-datetime": "^1.4.1",
     "chai-string": "^1.3.0",
     "chai-subset": "^1.4.0",
-    "eslint": "^3.12.2",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^17.1.0",
     "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-promise": "^3.3.0",
     "eslint-plugin-standard": "^2.0.1",
     "express": "^4.16.4",
