@@ -1,32 +1,30 @@
-var express = require('express')
-var path = require('path')
+const express = require('express');
+const path = require('path');
 
-var app = express()
+const app = express();
 
-var Boleto = require('../index').Boleto
+const { Boleto } = require('../index');
 
-var boleto = new Boleto({
-  'banco': 'santander',
-  'data_emissao': new Date(),
-  'data_vencimento': new Date(new Date().getTime() + 5 * 24 * 3600 * 1000),
-  'valor': 1500,
-  'nosso_numero': '1234567',
-  'numero_documento': '123123',
-  'cedente': 'Pagar.me Pagamentos S/A',
-  'cedente_cnpj': '18727053000174',
-  'agencia': '3978',
-  'codigo_cedente': '6404154', // PSK (código da carteira)
-  'carteira': '102'
-})
+const boleto = new Boleto({
+  banco: 'santander',
+  data_emissao: new Date(),
+  data_vencimento: new Date(new Date().getTime() + 5 * 24 * 3600 * 1000),
+  valor: 1500,
+  nosso_numero: '1234567',
+  numero_documento: '123123',
+  cedente: 'Pagar.me Pagamentos S/A',
+  cedente_cnpj: '18727053000174',
+  agencia: '3978',
+  codigo_cedente: '6404154', // PSK (código da carteira)
+  carteira: '102',
+});
 
-console.log(boleto['linha_digitavel'])
+console.log(boleto.linha_digitavel);
 
-app.use(express.static(path.join(__dirname, '/../')))
+app.use(express.static(path.join(__dirname, '/../')));
 
-app.get('/', function (req, res) {
-  boleto.renderHTML(function (html) {
-    return res.send(html)
-  })
-})
+app.get('/', (req, res) => {
+  boleto.renderHTML(html => res.send(html));
+});
 
-app.listen(3003)
+app.listen(3003);
