@@ -1,108 +1,110 @@
-const ejs = require('ejs')
-const formatters = require('./formatters')
-const barcode = require('./barcode')
-const path = require('path')
-const moment = require('moment')
+/* eslint-disable no-param-reassign,no-bitwise */
+const ejs = require('ejs');
+const path = require('path');
+const moment = require('moment');
+const formatters = require('./formatters');
+const barcode = require('./barcode');
 
-var banks = null
+let banks = null;
 
-var hashString = function (string) {
-  var hash = 0
-  var i
-  var chr
-  var len
+const hashString = function (string) {
+  let hash = 0;
+  let i;
+  let chr;
+  let len;
 
-  if (string.length == 0) return hash
+  if (string.length === 0) return hash;
   for (i = 0, len = string.length; i < len; i++) {
-    chr = string.charCodeAt(i)
-    hash = ((hash << 5) - hash) + chr
-    hash |= 0 // Convert to 32bit integer
+    chr = string.charCodeAt(i);
+    hash = ((hash << 5) - hash) + chr;
+    hash |= 0; // Convert to 32bit integer
   }
-  return hash
-}
+  return hash;
+};
 
-var Boleto = function (options) {
+const Boleto = function (options) {
   if (!options) {
-    throw 'No options provided initializing Boleto.'
+    throw 'No options provided initializing Boleto.';
   }
 
-  this.bank = banks[options['banco']]
+  this.bank = banks[options.banco];
   if (!this.bank) {
-    throw 'Invalid bank.'
+    throw 'Invalid bank.';
   }
 
-  if (!options['data_emissao']) {
-    options['data_emissao'] = moment().utc()
+  if (!options.data_emissao) {
+    options.data_emissao = moment().utc();
   } else {
-    options['data_emissao'] = moment(moment(options['data_emissao']).utc().format('YYYY-MM-DD'))
+    options.data_emissao = moment(moment(options.data_emissao).utc().format('YYYY-MM-DD'));
   }
 
-  if (!options['data_vencimento']) {
-    options['data_vencimento'] = moment().utc().add('5', 'days')
+  if (!options.data_vencimento) {
+    options.data_vencimento = moment().utc().add('5', 'days');
   } else {
-    options['data_vencimento'] = moment(moment(options['data_vencimento']).utc().format('YYYY-MM-DD'))
+    options.data_vencimento = moment(moment(options.data_vencimento).utc().format('YYYY-MM-DD'));
   }
 
-  for (var key in options) {
-    this[key] = options[key]
+  const t = this;
+  Object.keys(options).forEach((k) => {
+    t[k] = options[k];
+  });
+
+  this.pagador = formatters.htmlString(this.pagador);
+  this.instrucoes = formatters.htmlString(this.instrucoes);
+
+  if (!this.local_de_pagamento) {
+    this.local_de_pagamento = `Até o vencimento, preferencialmente no Banco ${formatters.capitalize(this.banco)}`;
   }
 
-  this['pagador'] = formatters.htmlString(this['pagador'])
-  this['instrucoes'] = formatters.htmlString(this['instrucoes'])
+  this._calculate();
+};
 
-  if (!this['local_de_pagamento']) {
-    this['local_de_pagamento'] = 'Até o vencimento, preferencialmente no Banco ' + formatters.capitalize(this['banco'])
+Boleto.barcodeRenderEngine = 'img';
+
+
+Boleto.prototype._calculate = function () {
+  this.codigo_banco = `${this.bank.options.codigo}-${formatters.mod11(this.bank.options.codigo)}`;
+  if (!this.nosso_numero_dv) {
+    this.nosso_numero_dv = formatters.mod11(this.nosso_numero.toString());
   }
-
-  this._calculate()
-}
-
-Boleto.barcodeRenderEngine = 'img'
-
-
-Boleto.prototype._calculate = function() {
-  this['codigo_banco'] = this.bank.options.codigo + "-" + formatters.mod11(this.bank.options.codigo);
-  if(!this['nosso_numero_dv']) {
-    this['nosso_numero_dv'] = formatters.mod11(this['nosso_numero'].toString());
-  }
-  this['barcode_data'] = this.bank.barcodeData(this);
-  this['linha_digitavel'] = this.bank.linhaDigitavel(this['barcode_data']);
-}
+  this.barcode_data = this.bank.barcodeData(this);
+  this.linha_digitavel = this.bank.linhaDigitavel(this.barcode_data);
+};
 
 Boleto.prototype.renderHTML = function (callback) {
-  var self = this
+  const self = this;
 
-  var renderOptions = self.bank.options
-  renderOptions.boleto = self
+  const renderOptions = self.bank.options;
+  renderOptions.boleto = self;
 
   // Copy renderHelper's methods to renderOptions
-  for (var key in formatters) {
-    renderOptions[key] = formatters[key]
+  Object.keys(formatters).forEach((k) => {
+    renderOptions[k] = formatters[k];
+  });
+
+  renderOptions.barcode_render_engine = Boleto.barcodeRenderEngine;
+  renderOptions.barcode_height = '50';
+
+  if (Boleto.barcodeRenderEngine === 'bmp') {
+    renderOptions.barcode_data = barcode.bmpLineForBarcodeData(self.barcode_data);
+  } else if (Boleto.barcodeRenderEngine === 'img') {
+    renderOptions.barcode_data = barcode.binaryRepresentationForBarcodeData(self.barcode_data);
   }
 
-  renderOptions['barcode_render_engine'] = Boleto.barcodeRenderEngine
-  renderOptions['barcode_height'] = '50'
-
-  if (Boleto.barcodeRenderEngine == 'bmp') {
-    renderOptions['barcode_data'] = barcode.bmpLineForBarcodeData(self['barcode_data'])
-  } else if (Boleto.barcodeRenderEngine == 'img') {
-    renderOptions['barcode_data'] = barcode.binaryRepresentationForBarcodeData(self['barcode_data'])
-  }
-
-  renderOptions['boleto']['linha_digitavel_hash'] = hashString(renderOptions['boleto']['linha_digitavel']).toString()
+  renderOptions.boleto.linha_digitavel_hash = hashString(renderOptions.boleto.linha_digitavel).toString();
 
   ejs.renderFile(path.join(__dirname, '/../assets/layout.ejs'), renderOptions, {
-    cache: true
-  }, function (err, html) {
+    cache: true,
+  }, (err, html) => {
     if (err) {
-      throw new Error(err)
+      throw new Error(err);
     }
 
-    callback(html)
-  })
-}
+    callback(html);
+  });
+};
 
 module.exports = function (_banks) {
-  banks = _banks
-  return Boleto
-}
+  banks = _banks;
+  return Boleto;
+};
