@@ -1,6 +1,5 @@
-var crypto = require('crypto')
+const crypto = require('crypto');
 
 exports.calculateLineChecksum = function (line) {
-  return crypto.createHash('sha1').update(line).digest('hex')
-}
-
+  return crypto.createHash('sha1').update(line).digest('hex');
+};
