@@ -1,7 +1,7 @@
 module.exports = function (banks) {
   return {
-    parse: function (bankName, fileContent) {
-      return banks[bankName].parseEDIFile(fileContent)
-    }
-  }
-}
+    parse(bankName, fileContent) {
+      return banks[bankName].parseEDIFile(fileContent);
+    },
+  };
+};
