@@ -1,102 +1,102 @@
-const escapeXML = require('ejs').escapeXML
-const moment = require('moment')
+/* eslint-disable no-param-reassign */
+const escapeXML = require('ejs').escapeXML;
+const moment = require('moment');
 
 exports.capitalize = function (string) {
-  return string.charAt(0).toUpperCase() + string.slice(1)
-}
+  return string.charAt(0).toUpperCase() + string.slice(1);
+};
 
 exports.addTrailingZeros = function (string, length) {
-  string = string.toString()
+  string = string.toString();
 
   while (string.length < length) {
-    string = '0' + string
+    string = `0${string}`;
   }
 
-  return string
-}
+  return string;
+};
 
 exports.formatAmount = function (amount) {
-  amount = amount.toString()
-  var cents = exports.addTrailingZeros(amount.substring(amount.length - 2, amount.length), 2)
-  var integers = exports.addTrailingZeros(amount.substring(0, amount.length - 2), 1)
+  amount = amount.toString();
+  const cents = exports.addTrailingZeros(amount.substring(amount.length - 2, amount.length), 2);
+  const integers = exports.addTrailingZeros(amount.substring(0, amount.length - 2), 1);
 
-  var newIntegers = ''
+  let newIntegers = '';
 
-  for (var i = 0; i < integers.length; i++) {
-    if (i > 0 && (integers.length - i) % 3 == 0) newIntegers += '.'
-    newIntegers += integers[i]
+  for (let i = 0; i < integers.length; i++) {
+    if (i > 0 && (integers.length - i) % 3 === 0) newIntegers += '.';
+    newIntegers += integers[i];
   }
 
-  return 'R$ ' + newIntegers + ',' + cents
-}
+  return `R$ ${newIntegers},${cents}`;
+};
 
 exports.formatDate = function (date) {
-  return moment(date).format('DD/MM/YYYY')
-}
+  return moment(date).format('DD/MM/YYYY');
+};
 
 exports.mod11 = function (num, base, r) {
-  if (!base) base = 9
-  if (!r) r = 0
+  if (!base) base = 9;
+  if (!r) r = 0;
 
-  var soma = 0
-  var fator = 2
+  let soma = 0;
+  let fator = 2;
 
-  for (var i = num.length - 1; i >= 0; i--) {
-    var parcial = parseInt(num[i]) * fator
-    soma += parcial
+  for (let i = num.length - 1; i >= 0; i--) {
+    soma += parseInt(num[i], 10) * fator;
 
-    if (fator == base) {
-      fator = 1
+    if (fator === base) {
+      fator = 1;
     }
 
-    fator++
+    fator += 1;
   }
 
-  if (r == 0) {
-    soma *= 10
-    var digito = soma % 11
-    return digito == 10 ? 0 : digito
-  } else if (r == 1) {
-    return soma % 11
+  if (r === 0) {
+    soma *= 10;
+    const digito = soma % 11;
+    return digito === 10 ? 0 : digito;
+  } if (r === 1) {
+    return soma % 11;
   }
-}
+};
 
 exports.mod10 = function (num) {
-  var total = 0
-  var fator = 2
+  let total = 0;
+  let fator = 2;
 
-  for (var i = num.length - 1; i >= 0; i--) {
-    var temp = (parseInt(num[i]) * fator).toString()
-    var tempSum = 0
-    for (var j = 0; j < temp.length; j++) {
-      tempSum += parseInt(temp[j])
+  for (let i = num.length - 1; i >= 0; i--) {
+    const temp = (parseInt(num[i], 10) * fator).toString();
+    let tempSum = 0;
+    for (let j = 0; j < temp.length; j++) {
+      tempSum += parseInt(temp[j], 10);
     }
-    total += tempSum
-    fator = (fator == 2) ? 1 : 2
+    total += tempSum;
+    fator = (fator === 2) ? 1 : 2;
   }
 
-  var resto = total % 10
-  return (resto == 0) ? 0 : (10 - resto)
-}
+  const resto = total % 10;
+  return (resto === 0) ? 0 : (10 - resto);
+};
 
 exports.fatorVencimento = function (date) {
-  const parsedDate = moment(date).utc().format('YYYY-MM-DD')
-  const startDate = moment('1997-10-07').utc().format('YYYY-MM-DD')
-  return exports.addTrailingZeros(moment(parsedDate).diff(startDate, 'days'), 4)
-}
+  const parsedDate = moment(date).utc().format('YYYY-MM-DD');
+  const startDate = moment('1997-10-07').utc().format('YYYY-MM-DD');
+  return exports.addTrailingZeros(moment(parsedDate).diff(startDate, 'days'), 4);
+};
 
 exports.dateFromEdiDate = function (ediDate) {
-  return new Date(parseInt(ediDate.substring(4, 8)), parseInt(ediDate.substring(2, 4)) - 1, parseInt(ediDate.substring(0, 2)))
-}
+  return new Date(parseInt(ediDate.substring(4, 8), 10), parseInt(ediDate.substring(2, 4), 10) - 1, parseInt(ediDate.substring(0, 2), 10));
+};
 
 exports.removeTrailingZeros = function (string) {
-  while (string.charAt(0) == '0') {
-    string = string.substring(1, string.length)
+  while (string.charAt(0) === '0') {
+    string = string.substring(1, string.length);
   }
 
-  return string
-}
+  return string;
+};
 
 exports.htmlString = function (str) {
-  return str ? escapeXML(str).replace(/\n/g, '<br/>') : str
-}
+  return str ? escapeXML(str).replace(/\n/g, '<br/>') : str;
+};
